Move checkGenerator and co to Test.Util.QuickCheck
